refactor(odata2model): extract generating stuff into own generator classes
