+++ packages/odata2model/src/generator/GeneratorModel.ts
@@ -1,0 +1,7 @@
+import { SourceFile } from "ts-morph";
+
+import { DataModel } from "./../data-model/DataModel";
+
+export interface TsGenerator {
+  generate(dataModel: DataModel, sourceFile: SourceFile): void;
+}
